carton: use VLQ format for string ids

+++ carton/src/string_table.rs
@@ -4,32 +4,28 @@ use crate::stream::Encode;
 use crate::stream::writing::write_string;
 
 /// Data structure that stores all strings used in a carton. Any strings outside of the string table are meant to be
-/// referenced using their corresponding string table ID. String table IDs are within the range `0..2**31`.
+/// referenced using their corresponding string table ID. String table IDs are within the range `0..2**60`.
 ///
-/// Strings can have a length within a range `0..2**16`.
+/// Strings can have a length within a range `0..2**60`.
 ///
-/// String table IDs are encoded using a gimmicky format. IDs less than 2**15 are encoded using two bytes, with the
-/// MSB set to 0. IDs greater than or equal to 2**15 are encoded using four bytes, with the 16th bit set to 1. String
-/// IDs within `0..2**15` save two bytes, reducing storage requirements.
-///
-/// TODO is this encoding a good idea? Does it actually save space?
+/// String table IDs are written using a variable length format with a 2 byte granularity.
 #[derive(Debug, Default)]
 pub struct StringTable {
-	mapping: HashMap<String, u32>,
-	sorted_mapping: BTreeMap<u32, String>,
+	mapping: HashMap<String, u64>,
+	sorted_mapping: BTreeMap<u64, String>,
 }
 
 impl StringTable {
 	/// Insert a string into the string table. Duplicates cannot be inserted.
-	pub fn insert(&mut self, string: &str) -> u32 {
-		let index = self.mapping.len() as u32;
+	pub fn insert(&mut self, string: &str) -> u64 {
+		let index = self.mapping.len() as u64;
 		self.mapping.insert(String::from(string), index);
 		self.sorted_mapping.insert(index, String::from(string));
 		return index;
 	}
 
 	/// Get a string's ID from the string table.
-	pub fn get(&self, string: &str) -> Option<u32> {
+	pub fn get(&self, string: &str) -> Option<u64> {
 		if let Some(index) = self.mapping.get(string) {
 			Some(*index)
 		} else {
